Rename module Absithe.Plug, support multiple ops
