+++ test/support/test_schema.ex
@@ -1,0 +1,53 @@
+defmodule Absinthe.Plug.TestSchema do
+  use Absinthe.Schema
+  alias Absinthe.Type
+
+  # Example data
+  @items %{
+    "foo" => %{id: "foo", name: "Foo"},
+    "bar" => %{id: "bar", name: "Bar"}
+  }
+
+  def query do
+    %Type.Object{
+      fields: fields(
+        item: [
+          type: :item,
+          args: args(
+            id: [type: non_null(:string)]
+          ),
+          resolve: fn %{id: item_id}, _ ->
+            {:ok, @items[item_id]}
+          end
+        ]
+      )
+    }
+  end
+
+  def mutation do
+    %Type.Object{
+      fields: fields(
+        addItem: [
+          type: :item,
+          args: args(
+            name: [
+              type: :string
+            ]
+          )
+        ]
+      )
+    }
+  end
+
+  @absinthe :type
+  def item do
+    %Type.Object{
+      description: "An item",
+      fields: fields(
+        id: [type: :id],
+        name: [type: :string]
+      )
+    }
+  end
+
+end
